AbstractDeclarativeProvider: cache provided code
Add a Cache<K, V> class and use it in AbstractDeclarativeProvider
to cache the return values of getFunctionsFor() and getProcsFor().
This is faster and avoids the previous problem where on each call
to these methods, the collect* methods were called again, adding
the created EObjects AGAIN to the providerResource (resulting in
a lot of duplicates).

Also, change the implementation of StdlibProvider (override the
collect*, not the get* methods) to take advantage of caching in
the super class.