heavy changes to variable declaration syntax
If 'const' is specified, 'var' is redundant. In general,
if a type is specified, it is now written after the name,
separated by a colon. To be unambigous, add new terminal
COLON. Even if a type is specified, either 'var' or 'const'
must be present.

Exceptions for parameters (just name : type), matrix init
counters.